Improvements in the energy-aware scheduler.
The energy aware scheduler now supports 3 different modes - "immediately",
"last-minute" and "energy-aware".
A simple approach is used to collect energy prices if the scheduler is
in "energy-aware" mode. It sends a request to a specific endpoint, in case the
endpoint is not accessible then the scheduler will start the workflow immediately.